Drop the one-shot v1->v3 migration script
The bulk migration has been applied to the live fiber directories, and
VC3D upgrades any straggler v1 file on save, so the standalone script
and its shared loader-checks module are no longer needed. The golden
lasagna-segment fixture stays: it pins the default descriptor bytes that
external tooling (fiber_merge validators) depends on.
